Add FrOS kernel
The FrOS kernel is a ready made x86 kernel we can use for further development of DustOS.
I removed the main Makefile due to issues with the GRUB bootloader.
This will be fixed at somepoint.
You can still build and run the kernel by going to src/kernel and running make.

Signed-off-by: Harry Fellowes <harryfellowes1@gmail.com>